requirements.txt
astroid==2.2.5
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.0.0
pylint==2.3.1
pylint-flask==0.6
pylint-plugin-utils==0.5
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.1
virtualenv==16.5.0
visitor==0.1.3
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
